moved ng-zorro-antd base fields
