+++ client/src/app/forms/config/ng-zorro-antd/textarea/src/textarea.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { FormlyModule } from '@ngx-formly/core';
 import { ReactiveFormsModule } from '@angular/forms';
 
-import { FormlyNzFormFieldModule } from '@libs/ng-zorro-antd/form-field';
+import { FormlyNzFormFieldModule } from '../../form-field';
 import { NzInputModule } from 'ng-zorro-antd/input';
 
 import { FormlyFieldTextArea } from './textarea.type';
