+++ client/src/app/forms/config/types/source-array/source-array.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { FormlyModule } from '@ngx-formly/core';
 import { FormlySelectModule } from '@ngx-formly/core/select';
-import { FormlyNzFormFieldModule } from '@libs/ng-zorro-antd/form-field';
+import { FormlyNzFormFieldModule } from '@forms/config/ng-zorro-antd/form-field';
 import { NzSelectModule } from 'ng-zorro-antd/select';
 import { CvcSourceInputTypeModule } from '../source-input/source-input.module';
 import { CvcMultiFieldTypeModule } from '../multi-field/multi-field.module';
